+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Devop
 
 * Update home features [#2935](https://github.com/MyEtherWallet/MyEtherWallet/pull/2935)
+* Add MEWconnect [#2934](https://github.com/MyEtherWallet/MyEtherWallet/pull/2934)
+* Fix send and sign tx [#2929](https://github.com/MyEtherWallet/MyEtherWallet/pull/2929)
 * Merge UI fixes [#2916](https://github.com/MyEtherWallet/MyEtherWallet/pull/2916)
 * Fix create wallet nnemonic layout [#2912](https://github.com/MyEtherWallet/MyEtherWallet/pull/2912)
 * Fix create wallet keystore layout [#2911](https://github.com/MyEtherWallet/MyEtherWallet/pull/2911)
@@ -43,3 +45,4 @@
 * Apply MEW components to dashboard [#2378](https://github.com/MyEtherWallet/MyEtherWallet/pull/2378)
 * Add MY TOKENS VALUE component for Dashboard [#2015](https://github.com/MyEtherWallet/MyEtherWallet/pull/2015)
 * Update my eth balance component chart [#1987](https://github.com/MyEtherWallet/MyEtherWallet/pull/1987)
+* Sign and Verify Message [#2841](https://github.com/MyEtherWallet/MyEtherWallet/pull/2841)